[weather.gismeteo] 0.6.2
