+++ weather.gismeteo/resources/language/resource.language.is_is/strings.po
@@ -1,0 +1,269 @@
+# Kodi Media Center language file
+# Addon Name: Gismeteo
+# Addon id: weather.gismeteo
+# Addon Provider: vl.maksime
+
+# Íslensk þýðing: Sveinn í Felli
+
+msgid ""
+msgstr ""
+
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: is\n"
+
+# Categories
+msgctxt "#32101"
+msgid "Location setup"
+msgstr "Stilling staðsetningar"
+
+msgctxt "#32102"
+msgid "Advanced"
+msgstr "Ítarlegt"
+
+msgctxt "#32103"
+msgid "Expert"
+msgstr "Sérfræðingar"
+
+# Location setup
+msgctxt "#32110"
+msgid "Current Location"
+msgstr "Núverandi staðsetning"
+
+msgctxt "#32111"
+msgid "Location 1"
+msgstr "Staðsetning 1"
+
+msgctxt "#32112"
+msgid "Location 2"
+msgstr "Staðsetning 2"
+
+msgctxt "#32113"
+msgid "Location 3"
+msgstr "Staðsetning 3"
+
+msgctxt "#32114"
+msgid "Location 4"
+msgstr "Staðsetning 4"
+
+msgctxt "#32115"
+msgid "Location 5"
+msgstr "Staðsetning 5"
+
+msgctxt "#32116"
+msgid "Location 1 ID"
+msgstr "Auðkenni staðsetningar 1"
+
+msgctxt "#32117"
+msgid "Location 2 ID"
+msgstr "Auðkenni staðsetningar 2"
+
+msgctxt "#32118"
+msgid "Location 3 ID"
+msgstr "Auðkenni staðsetningar 3"
+
+msgctxt "#32119"
+msgid "Location 4 ID"
+msgstr "Auðkenni staðsetningar 4"
+
+msgctxt "#32120"
+msgid "Location 5 ID"
+msgstr "Auðkenni staðsetningar 5"
+
+# Advanced
+msgctxt "#32122"
+msgid "Weekend"
+msgstr "Helgi"
+
+msgctxt "#32123"
+msgid "Saturday/Sunday"
+msgstr "Laugardag/Sunnudag"
+
+msgctxt "#32124"
+msgid "Friday/Saturday"
+msgstr "Föstudag/Laugardag"
+
+msgctxt "#32125"
+msgid "Thursday/Friday"
+msgstr "Fimmtudag/Föstudag"
+
+msgctxt "#32130"
+msgid "Time zone"
+msgstr "Tímabelti"
+
+msgctxt "#32131"
+msgid "from system settings"
+msgstr "úr stillingum kerfis"
+
+msgctxt "#32132"
+msgid "from location"
+msgstr "út frá staðsetningu"
+
+msgctxt "#32133"
+msgid "Measure units"
+msgstr "Mælieiningar"
+
+msgctxt "#32134"
+msgid "Pressure"
+msgstr "Þrýstingur"
+
+msgctxt "#32135"
+msgid "Precipitation"
+msgstr "Úrkoma"
+
+msgctxt "#32136"
+msgid "Clear cache"
+msgstr "Hreinsa skyndiminni"
+
+msgctxt "#32200"
+msgid "User interface language"
+msgstr "Tungumál notandaviðmóts"
+
+msgctxt "#32201"
+msgid "Russian"
+msgstr "Rússneska"
+
+msgctxt "#32202"
+msgid "Ukrainian"
+msgstr "Úkraníska"
+
+msgctxt "#32203"
+msgid "Lithuanian"
+msgstr "Litháíska"
+
+msgctxt "#32204"
+msgid "Latvian"
+msgstr "Lettneska"
+
+msgctxt "#32205"
+msgid "English"
+msgstr "Enska"
+
+msgctxt "#32206"
+msgid "Romanian"
+msgstr "Rúmenska"
+
+msgctxt "#32207"
+msgid "German"
+msgstr "Þýska"
+
+msgctxt "#32208"
+msgid "Polish"
+msgstr "Pólska"
+
+msgctxt "#32220"
+msgid "Gismeteo icons"
+msgstr "Gismeteo-tákn"
+
+msgctxt "#32221"
+msgid "Use Gismeteo icons"
+msgstr "Nota Gismeteo-tákn"
+
+msgctxt "#32222"
+msgid "Install icon package"
+msgstr "Setja inn táknmyndapakka"
+
+msgctxt "#32223"
+msgid "Requires Kodi 16+ and skin support"
+msgstr "Krefst Kodi 16+ og stuðnings við skinn"
+
+msgctxt "#32299"
+msgid "Language"
+msgstr "Tungumál"
+
+msgctxt "#32301"
+msgid "a/p"
+msgstr "a/p"
+
+# Speed
+msgctxt "#32401"
+msgid "m/s"
+msgstr "m/sek"
+
+msgctxt "#32402"
+msgid "km/h"
+msgstr "km/klst"
+
+msgctxt "#32403"
+msgid "m/min"
+msgstr "m/mín"
+
+msgctxt "#32404"
+msgid "ft/h"
+msgstr "ft/klst"
+
+msgctxt "#32405"
+msgid "ft/min"
+msgstr "ft/mín"
+
+msgctxt "#32406"
+msgid "ft/s"
+msgstr "ft/sek"
+
+msgctxt "#32407"
+msgid "knots"
+msgstr "hnútar"
+
+msgctxt "#32408"
+msgid "Beaufort"
+msgstr "Beaufort"
+
+msgctxt "#32409"
+msgid "inch/s"
+msgstr "tomma/sek"
+
+msgctxt "#32410"
+msgid "yard/s"
+msgstr "yard/sek"
+
+msgctxt "#32411"
+msgid "Furlong/Fortnight"
+msgstr "Furlong/Fortnight"
+
+msgctxt "#32412"
+msgid "mph"
+msgstr "mph"
+
+msgctxt "#32413"
+msgid "kts"
+msgstr "kts"
+
+# Wind direction
+msgctxt "#32420"
+msgid "calm"
+msgstr "logn"
+
+# Precipitation
+msgctxt "#32430"
+msgid "mm"
+msgstr "mm"
+
+msgctxt "#32431"
+msgid "inch"
+msgstr "tommur"
+
+msgctxt "#32432"
+msgid "n/a"
+msgstr "ekki tiltækt"
+
+# Pressure
+msgctxt "#32440"
+msgid "mmHg"
+msgstr "mmHg"
+
+msgctxt "#32441"
+msgid "hPa"
+msgstr "hPa"
+
+msgctxt "#32442"
+msgid "mbar"
+msgstr "mbar"
+
+msgctxt "#32443"
+msgid "inHg"
+msgstr "inHg"
+
+# Messages
+msgctxt "#32501"
+msgid "Connection error"
+msgstr "Tengingarvilla"
