+++ weather.gismeteo/resources/language/resource.language.sv_se/strings.po
@@ -1,0 +1,275 @@
+# Kodi Media Center language file
+# Addon Name: Gismeteo
+# Addon id: weather.gismeteo
+# Addon Provider: vl.maksime
+
+msgid ""
+msgstr ""
+
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: sv\n"
+
+# Categories
+
+msgctxt "#32101"
+msgid "Location setup"
+msgstr "Platsinställning"
+
+msgctxt "#32102"
+msgid "Advanced"
+msgstr "Avancerat"
+
+msgctxt "#32103"
+msgid "Expert"
+msgstr "Expert"
+
+# Location setup
+
+msgctxt "#32110"
+msgid "Current Location"
+msgstr "Nuvarande plats"
+
+msgctxt "#32111"
+msgid "Location 1"
+msgstr "Plats 1"
+
+msgctxt "#32112"
+msgid "Location 2"
+msgstr "Plats 2"
+
+msgctxt "#32113"
+msgid "Location 3"
+msgstr "Plats 3"
+
+msgctxt "#32114"
+msgid "Location 4"
+msgstr "Plats 4"
+
+msgctxt "#32115"
+msgid "Location 5"
+msgstr "Plats 5"
+
+msgctxt "#32116"
+msgid "Location 1 ID"
+msgstr "ID för plats 1"
+
+msgctxt "#32117"
+msgid "Location 2 ID"
+msgstr "ID för plats 2"
+
+msgctxt "#32118"
+msgid "Location 3 ID"
+msgstr "ID för plats 3"
+
+msgctxt "#32119"
+msgid "Location 4 ID"
+msgstr "ID för plats 4"
+
+msgctxt "#32120"
+msgid "Location 5 ID"
+msgstr "ID för plats 5"
+
+# Advanced
+
+msgctxt "#32122"
+msgid "Weekend"
+msgstr "Veckoslut"
+
+msgctxt "#32123"
+msgid "Saturday/Sunday"
+msgstr "Lördag/Söndag"
+
+msgctxt "#32124"
+msgid "Friday/Saturday"
+msgstr "Fredag/Lördag"
+
+msgctxt "#32125"
+msgid "Thursday/Friday"
+msgstr "Torsdag/Fredag"
+
+msgctxt "#32130"
+msgid "Time zone"
+msgstr "Tidszon"
+
+msgctxt "#32131"
+msgid "from system settings"
+msgstr "från systeminställningarna"
+
+msgctxt "#32132"
+msgid "from location"
+msgstr "från plats"
+
+msgctxt "#32133"
+msgid "Measure units"
+msgstr "Mätenheter"
+
+msgctxt "#32134"
+msgid "Pressure"
+msgstr "Tryck"
+
+msgctxt "#32135"
+msgid "Precipitation"
+msgstr "Nederbörd"
+
+msgctxt "#32136"
+msgid "Clear cache"
+msgstr "Rensa cache"
+
+msgctxt "#32200"
+msgid "User interface language"
+msgstr "Användargränssnittsspråk"
+
+msgctxt "#32201"
+msgid "Russian"
+msgstr "Ryska"
+
+msgctxt "#32202"
+msgid "Ukrainian"
+msgstr "Ukrainska"
+
+msgctxt "#32203"
+msgid "Lithuanian"
+msgstr "Litauiska"
+
+msgctxt "#32204"
+msgid "Latvian"
+msgstr "Lettiska"
+
+msgctxt "#32205"
+msgid "English"
+msgstr "Engelska"
+
+msgctxt "#32206"
+msgid "Romanian"
+msgstr "Rumänska"
+
+msgctxt "#32207"
+msgid "German"
+msgstr "Tyska"
+
+msgctxt "#32208"
+msgid "Polish"
+msgstr "Polska"
+
+msgctxt "#32220"
+msgid "Gismeteo icons"
+msgstr "Gismeteo-ikoner"
+
+msgctxt "#32221"
+msgid "Use Gismeteo icons"
+msgstr "Använd Gismeteo-ikoner"
+
+msgctxt "#32222"
+msgid "Install icon package"
+msgstr "Installera från ikonpaket"
+
+msgctxt "#32223"
+msgid "Requires Kodi 16+ and skin support"
+msgstr "Kräver Kodi 16+ och skalstöd"
+
+msgctxt "#32299"
+msgid "Language"
+msgstr "Språk"
+
+msgctxt "#32301"
+msgid "a/p"
+msgstr "a/p"
+
+# Speed
+
+msgctxt "#32401"
+msgid "m/s"
+msgstr "m/s"
+
+msgctxt "#32402"
+msgid "km/h"
+msgstr "km/h"
+
+msgctxt "#32403"
+msgid "m/min"
+msgstr "m/min"
+
+msgctxt "#32404"
+msgid "ft/h"
+msgstr "ft/h"
+
+msgctxt "#32405"
+msgid "ft/min"
+msgstr "ft/min"
+
+msgctxt "#32406"
+msgid "ft/s"
+msgstr "ft/s"
+
+msgctxt "#32407"
+msgid "knots"
+msgstr "knop"
+
+msgctxt "#32408"
+msgid "Beaufort"
+msgstr "Beaufort"
+
+msgctxt "#32409"
+msgid "inch/s"
+msgstr "tum/s"
+
+msgctxt "#32410"
+msgid "yard/s"
+msgstr "yard/s"
+
+msgctxt "#32411"
+msgid "Furlong/Fortnight"
+msgstr "Furlong/Fortnight"
+
+msgctxt "#32412"
+msgid "mph"
+msgstr "mph"
+
+msgctxt "#32413"
+msgid "kts"
+msgstr "kts"
+
+# Wind direction
+
+msgctxt "#32420"
+msgid "calm"
+msgstr "lugnt"
+
+# Precipitation
+
+msgctxt "#32430"
+msgid "mm"
+msgstr "mm"
+
+msgctxt "#32431"
+msgid "inch"
+msgstr "tum"
+
+msgctxt "#32432"
+msgid "n/a"
+msgstr "n/a"
+
+# Pressure
+
+msgctxt "#32440"
+msgid "mmHg"
+msgstr "mmHg"
+
+msgctxt "#32441"
+msgid "hPa"
+msgstr "hPa"
+
+msgctxt "#32442"
+msgid "mbar"
+msgstr "mbar"
+
+msgctxt "#32443"
+msgid "inHg"
+msgstr "inHg"
+
+# Messages
+
+msgctxt "#32501"
+msgid "Connection error"
+msgstr "Anslutningsfel"
