+++ weather.gismeteo/resources/libs/simpleweather/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+# License: GPL v.3 https://www.gnu.org/copyleft/gpl.html
+
+from __future__ import unicode_literals
+
+from .simpleweather import (WeatherProperties, Weather,
+                            Addon, py2_encode, py2_decode)
+from .webclient import WebClient, WebClientError
+
+__all__ = ['WeatherProperties', 'Weather',
+           'Addon', 'py2_encode', 'py2_decode',
+           'WebClient', 'WebClientError']
